modified files to match makis python api
