+++ core/fokia/provisioner.py
@@ -9,6 +9,7 @@ from kamaki.clients import astakos, cyclades
 from kamaki.clients import ClientError
 from kamaki.clients.utils import https
 from kamaki.cli.config import Config as KamakiConfig
+from fokia.cluster_error_constants import *
 
 # TODO: remove this and actually use ssl cert files
 https.patch_ignore_ssl()
@@ -45,6 +46,15 @@ class Provisioner:
             cyclades.CycladesComputeClient.service_type)
         logger.info("Initiating Cyclades client")
         self.cyclades = cyclades.CycladesComputeClient(compute_url, auth_token)
+
+        # Create the network client
+        networkURL = self.astakos.get_endpoint_url(
+            cyclades.CycladesNetworkClient.service_type)
+        self.network_client = cyclades.CycladesNetworkClient(networkURL, auth_token)
+
+        # Constants
+        self.Bytes_to_GB = 1024*1024*1024
+        self.Bytes_to_MB = 1024*1024
 
     def find_flavor(self, **kwargs):
         """
@@ -109,6 +119,82 @@ class Provisioner:
         except ClientError as ex:
             raise ex
         return okeanos_response
+
+    def get_quotas(self, **kwargs):
+        """
+        Get the user quotas for the defined project.
+        :return: user quotas object
+        """
+        return self.astakos.get_quotas()
+
+    def check_all_resources(self, quotas, **kwargs):
+        """
+        Checks user's quota for every requested resource.
+        Returns True if everything available.
+        :param **kwargs: arguments
+        """
+        project_id = self.find_project_id(**kwargs)['id']
+        # quotas = self.get_quotas()
+
+        # Check for VMs
+        pending_vm = quotas[project_id]['cyclades.vm']['pending']
+        limit_vm = quotas[project_id]['cyclades.vm']['limit']
+        usage_vm = quotas[project_id]['cyclades.vm']['usage']
+        available_vm = limit_vm - usage_vm - pending_vm
+        if available_vm < kwargs['cluster_size']:
+            msg = 'Cyclades VMs out of limit'
+            raise ClientError(msg, error_quotas_cluster_size)
+            return False
+        # Check for CPUs
+        pending_cpu = quotas[project_id]['cyclades.cpu']['pending']
+        limit_cpu = quotas[project_id]['cyclades.cpu']['limit']
+        usage_cpu = quotas[project_id]['cyclades.cpu']['usage']
+        available_cpu = limit_cpu - usage_cpu - pending_cpu
+        if available_cpu < kwargs['vcpus']:
+            msg = 'Cyclades cpu out of limit'
+            raise ClientError(msg, error_quotas_cpu)
+            return False
+        # Check for RAM
+        pending_ram = quotas[project_id]['cyclades.ram']['pending']
+        limit_ram = quotas[project_id]['cyclades.ram']['limit']
+        usage_ram = quotas[project_id]['cyclades.ram']['usage']
+        available_ram = (limit_ram - usage_ram - pending_ram) / self.Bytes_to_MB
+        if available_ram < kwargs['ram']:
+            msg = 'Cyclades ram out of limit'
+            raise ClientError(msg, error_quotas_ram)
+            return False
+        # Check for Disk space
+        pending_cd = quotas[project_id]['cyclades.ram']['pending']
+        limit_cd = quotas[project_id]['cyclades.disk']['limit']
+        usage_cd = quotas[project_id]['cyclades.disk']['usage']
+        available_cyclades_disk_GB = (limit_cd - usage_cd - pending_cd) / self.Bytes_to_GB
+        if available_cyclades_disk_GB < kwargs['disk']:
+            msg = 'Cyclades disk out of limit'
+            raise ClientError(msg, error_quotas_cyclades_disk)
+            return False
+        # Check for public IPs
+        list_float_ips = self.network_client.list_floatingips()
+        pending_ips = quotas[project_id]['cyclades.floating_ip']['pending']
+        limit_ips = quotas[project_id]['cyclades.floating_ip']['limit']
+        usage_ips = quotas[project_id]['cyclades.floating_ip']['usage']
+        available_ips = limit_ips - usage_ips - pending_ips
+        for d in list_float_ips:
+            if d['instance_id'] is None and d['port_id'] is None:
+                available_ips += 1
+        if available_ips < kwargs['ip_request']:
+            msg = 'Public IPs out of limit'
+            raise ClientError(msg, error_quotas_cyclades_disk)
+            return False
+        # Check for networks
+        pending_net = quotas[project_id]['cyclades.network.private']['project_pending']
+        limit_net = quotas[project_id]['cyclades.network.private']['project_limit']
+        usage_net = quotas[project_id]['cyclades.network.private']['project_usage']
+        available_networks = limit_net - usage_net - pending_net
+        if available_networks < kwargs['network_request']:
+            msg = 'Private Network out of limit'
+            raise ClientError(msg, error_quotas_cyclades_disk)
+            return False
+        return True
 
 
 if __name__ == "__main__":
